Extract file watcher
