+++ src/engine/mod.rs
@@ -1,3 +1,4 @@
+pub mod filewatcher;
 mod http;
 pub mod runtime;
 mod stdlib;
